Fix interfaces failing to be set
Cause:
Interfaces are created with a command that is run in the background (line ends in '&'; in createNsVethPair). As a result, we do not wait until this procedure is done, and so the interface might not be available to modify by the time that we try to change its MAC address.

Fix:
Removed the '&'. We now wait for the add interface command to finish before configuring its MAC address.

Question:
Not clear on why the & was there to begin with. It is present in other locations in the source code, always in the last shell command (PipesExec) in a function